Recommit: Euler 1, 6, and 7
Euler 6 was in the wrong folder ("Euler 0003"). Removed it and placed it
in its correct folder.

Euler 1 and 7 are commented with problem descriptions, the solution,
side-notes, and of course, comments to explain the program.

Will continue this process for the rest of the Euler problems. Hopefully
the current comment decoration process is sustainable.